refactor: Rename some classes after rebase onto SQLiteDB branch
