+++ lib/build.gradle.kts
@@ -22,6 +22,7 @@ repositories {
 dependencies {
     api("org.apache.commons:commons-math3:3.6.1")
     implementation("com.google.guava:guava:31.1-jre")
+    implementation("org.jetbrains.kotlinx:kotlinx-serialization-json:1.5.0")
 
     implementation("com.google.code.gson:gson:2.8.5")
