+++ lib/src/main/kotlin/dbSave/DrawVertex.kt
@@ -1,4 +1,4 @@
-package treelib.controller
+package dbSave
 
 interface DrawVertex<Pack: Comparable<Pack>> {
     val value: Pack
